requirements.txt
streamlit
langchain
langchain-groq
langchain-community
langchain-experimental
sqlalchemy
pandas
matplotlib
seaborn
python-dotenv
faker